convert camera component to external class
