+++ Engine/src/camera/CameraControllerFreeLook.h
@@ -3,7 +3,7 @@
 #include <glm/glm.hpp>
 
 #include "camera/ICameraController.h"
-#include "component/Component.h"
+#include "component/CameraComponent.h"
 #include "core/Subscriber.h"
 
 class EngineAPI CameraControllerFreeLook : public ICameraController
