+++ Engine/src/serialize/Archiver.h
@@ -78,7 +78,7 @@ struct SerializedEntity
 	std::optional<PlayerController> playerController;
 	std::optional<MeshComponent> mesh;
 	std::optional<RenderableComponent> renderableComponent;
-	std::optional<CameraComponent> camera;
+	//std::optional<CameraComponent> camera;
 	std::optional<NativeScriptComponent> nsc;
 	std::optional<MaterialComponent> mat;
 	std::optional<DirectionalLight> dLight;
@@ -102,7 +102,7 @@ struct SerializedEntity
 		SERIALIZED_MEMBER(playerController);
 		SERIALIZED_MEMBER(mesh);
 		SERIALIZED_MEMBER(renderableComponent);
-		SERIALIZED_MEMBER(camera);
+		//SERIALIZED_MEMBER(camera);
 		SERIALIZED_MEMBER(nsc);
 		SERIALIZED_MEMBER(mat);
 		SERIALIZED_MEMBER(dLight);
